Module Fpu.Rename_all

Aliases floating point functions to their "constant" counterparts, including +., -., *. and /..

As described in the Fpu module documentation, there are problems when mixing some C-lib or ocaml native functions with interval programming on 64 bits machine.

The standard floating point functions results will always lie in the [low; high] interval computed by the Fpu module, but they are slightly different on 32 and 64 bits machines.

Using open Fpu.Rename_all at the beginning of your program guarantees that floating computation will give the same results on 32 and 64 bits machines. This is not mandatory but might help.

NB: while most transcendantal function are almost as fast, and sometimes faster than their "standard" ocaml counterparts, +., -., *. and /. are much slower (from 50% to 100% depending on the processor. If you want to rename transcendantal functions but not +., -., *. and /. then use the Fpu.Rename module.

val (+.) : float -> float -> float

Alias for Fpu.fadd.

val (-.) : float -> float -> float

Alias for Fpu.fsub.

val (*.) : float -> float -> float

Alias for Fpu.fmul.

val (/.) : float -> float -> float

Alias for Fpu.fdiv.

val mod_float : float -> float -> float

Alias for Fpu.fmod.

val exp : float -> float

Alias for Fpu.fexp.

val log : float -> float

Alias for Fpu.flog.

val (**) : float -> float -> float

Alias for Fpu.fpow.

val sin : float -> float

Alias for Fpu.fsin. Computes sin(x) for x ∈ [-2⁶³, 2⁶³].

val cos : float -> float

Alias for Fpu.fcos. Computes cos(x) for x ∈ [-2⁶³, 2⁶³].

val tan : float -> float

Alias for Fpu.ftan. Computes tan(x) for x ∈ [-2⁶³, 2⁶³].

val asin : float -> float

Alias for Fpu.fasin.

val acos : float -> float

Alias for Fpu.facos.

val atan : float -> float

Arc-tangent function using Fpu.fatan.

val atan2 : float -> float -> float

atan2 function using Fpu.fatan.

val cosh : float -> float

Alias for Fpu.fcosh.

val sinh : float -> float

Alias for Fpu.fsinh.

val tanh : float -> float

Alias for Fpu.ftanh.