Module Fpu.RoundUp

Functions rounding up their results.

include module type of Interval_base.RoundUp

Functions rounding up their results.

include Interval_base.DIRECTED with type t = float

type t = float

Type of numbers.

val zero : t

The neutral element for addition.

val one : t

The neutral element for multiplication.

val pi : t

Upper/lower bound on π.

val two_pi : t

Upper/lower bound on 2π.

val half_pi : t

Upper/lower bound on π/2.

val euler : t

Upper/lower bound on Euler's constant.

val float : int -> t

When t = float, the float function is exact on 32 bits machine but not on 64 bits machine with ints larger than 53 bits.

val dist : t -> t -> t

dist x y return the distance between x and y (i.e., |x-y|) rounded up or down according to the module.

val (+.) : t -> t -> t

val (-.) : t -> t -> t

val (*.) : t -> t -> t

val (/.) : t -> t -> t

val sqr : t -> t

sqr x returns an upper/lower bound on x².

val cbr : t -> t

cbr x returns an upper/lower bound on x³.

val pow_i : t -> int -> t

pow_i x n return a upper/lower bound on xⁿ.

val sqrt : t -> t

sqrt x return the square root of x rounded up/down.

val hypot : t -> t -> t

hypot x y returns sqrt(x *. x + y *. y) rounded up/down. Returns NaN if any of the arguments is NaN.

module U = Interval_base.I.U

Locally open to restore standard integer and floating point operators.

val (**) : float -> float -> float

x**y computes x at power y, rounded up, expanded to its mathematical limit when it exists.

val pow : float -> float -> float

pow x y computes x at power y, rounded up, for 0 < x < ∞ and -∞ < y < infty.

val exp : float -> float

Floating point exponential, rounded up.

val log : float -> float

Floating point log, rounded up.

val sin : float -> float

Computes sin(x) for x ∈ ]-2⁶³, 2⁶³[.

val cos : float -> float

Computes cos(x) for x in ]-2⁶³, 2⁶³[.

val tan : float -> float

Computes tan(x) for x in ]-2⁶³, 2⁶³[.

val asin : float -> float

Arc-sine function.

val acos : float -> float

Arc-cosine function.

val atan : float -> float

Arc-tangent function.

val atan2 : float -> float -> float

atan2 y x returns the arc tangent of y /. x. The signs of x and y are used to determine the quadrant of the result. Result is in radians and is between -π and π.

val sinh : float -> float

Computes the hyperbolic sine, sinh(x).

val cosh : float -> float

Computes the hyperbolic cosine, cosh(x).

val tanh : float -> float

Computes the hyperbolic tangent, tanh(x).