Module I.Arr

Operations on arrays of intervals.

val size_max : t array -> float

Computes the size of the largest interval of the interval vector.

val size_mean : t array -> float

Computes the mean of the size of intervals of the interval vector.

val to_string : ?fmt:( float -> 'b, 'a, 'b ) Stdlib.format -> t array -> string

to_string a returns a string representation of a.

parameter fmt
is the format used to print the two bounds of i. Default: "%g".

val pr : Stdlib.out_channel -> t array -> unit

Print the interval array to the channel. To be used with Printf format "%a".

val pp : Stdlib.Format.formatter -> t array -> unit

Print the interval array to the formatter. To be used with Format format "%a".

val fmt : ( float -> 'b, 'a, 'b ) Stdlib.format -> ( t array -> 'c, 'd, 'e, 'c ) Stdlib.format4